# Conflicts:
#	include/vcpkg/base/contractual-constants.h
#	include/vcpkg/spdx.h
#	include/vcpkg/vcpkglib.h
#	src/vcpkg/commands.build.cpp
#	src/vcpkg/commands.ci.cpp
#	src/vcpkg/commands.install.cpp
#	src/vcpkg/commands.set-installed.cpp
#	src/vcpkg/commands.upgrade.cpp
#	src/vcpkg/spdx.cpp
#	src/vcpkg/vcpkglib.cpp

+++ azure-pipelines/end-to-end-tests-dir/build-test-ports.ps1
@@ -28,6 +28,12 @@ Throw-IfFailed
 Run-Vcpkg @commonArgs --overlay-ports="$PSScriptRoot/../e2e-ports" --overlay-ports="$PSScriptRoot/../e2e-ports/broken-manifests" install control-file
 Throw-IfFailed
 
+$output = Run-VcpkgAndCaptureOutput @commonArgs --overlay-ports="$PSScriptRoot/../e2e-ports/broken-manifests" install broken-duplicate-overrides
+Throw-IfNotFailed
+if ($output -notmatch "vcpkg\.json: error: \$\.overrides\[1\] \(an override\): zlib already has an override") {
+    throw 'Did not detect duplicate override'
+}
+
 $output = Run-VcpkgAndCaptureOutput @commonArgs --overlay-ports="$PSScriptRoot/../e2e-ports/broken-manifests" install broken-no-name
 Throw-IfNotFailed
 if ($output -notmatch "missing required field 'name'") {
